6: Fix video & coco export
