+++ src/encord_active/lib/metrics/execute.py
@@ -10,8 +10,8 @@ from typing import Callable, List, Optional, Sequence, Tuple, Type, Union
 import faiss  # pylint: disable=unused-import
 from loguru import logger
 
+from encord_active.lib.common.data_utils import convert_image_bgr
 from encord_active.lib.common.iterator import DatasetIterator, Iterator
-from encord_active.lib.common.utils import convert_image_bgr
 from encord_active.lib.common.writer import StatisticsObserver
 from encord_active.lib.labels.classification import ClassificationType
 from encord_active.lib.labels.object import ObjectShape
